onboarding_tool/requirements.txt
anthropic>=0.20.0
pytrends>=4.9.2
reportlab>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
